Refactor: Moved cli package into seperate files
