+++ internal/cli/feedhandlers.go
@@ -1,0 +1,58 @@
+package cli
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/codingchem/gator/internal/database"
+	"github.com/codingchem/gator/internal/rss"
+	"github.com/google/uuid"
+)
+
+func handlerAgg(_ *state, _ command) error {
+	feed, err := rss.FetchFeed(context.Background(), "https://www.wagslane.dev/index.xml")
+	if err != nil {
+		return err
+	}
+	fmt.Println(feed)
+	return nil
+}
+
+func handlerAddFeed(s *state, cmd command) error {
+	if len(cmd.args) != 2 {
+		return fmt.Errorf("Error: Addfeed requires 2 arguments!")
+	}
+	user, err := s.db.GetUser(context.Background(), s.config.CurrentUser)
+	if err != nil {
+		return err
+	}
+
+	_, err = s.db.CreateFeed(
+		context.Background(),
+		database.CreateFeedParams{
+			ID:        uuid.New(),
+			CreatedAt: time.Now(),
+			UpdatedAt: time.Now(),
+			Name:      cmd.args[1],
+			Url:       cmd.args[0],
+			UserID:    user.ID,
+		})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+func handlerListFeeds(s *state, cmd command) error {
+	if len(cmd.args) != 0 {
+		return fmt.Errorf("Error: feeds takes no arguments!")
+	}
+	feeds, err := s.db.GetFeeds(context.Background())
+	if err != nil {
+		return err
+	}
+	for _, feed := range feeds {
+		fmt.Printf("\nTitle: %s\nURL: %s\nCreated by: %s\n", feed.Name, feed.Url, feed.UserName)
+	}
+	return nil
+}
